Combine shipment service item stuff into one file
